Octaves, Persistence and Lacunarity
Added in octaves, persistence and lacunarity to affect the noise generated.

Octaves determine the amount of individual layers of noise, Persistence allows for the control of decrease in amplitude and Lacunarity allows for the control of increase in frequency.

Added seed and offset to allow for users to get the same result when using the same seed and for the user to offset on the x and y co-ords to get different noise but at the same level of detail.

Added OnValidate method to allow Unity to control where certain variables values can go, so that the program doesn't crash or have an unintended effect.

Half of width and height variables made so that noiseScale scales from the centre rather than from the top right.