Added ColorMap and Regions
Separated noiseMap generation from MapDsiplay (which now only displays the texture), which is now generated in TextureGenerator, as well as colorMap (noiseMap renamed to heightMap)

colorMap works in the same way as the heightMap does, by using the height (darkness of each pixel to determine height) of each pixel and assigning color to each element in an array. A struct has been created to allow users to create their own colours for the environment they want, with an example already filled in.

The TextureGenerator also makes use of a point filter mode and clamp texture wrap mode

An enum has been created to allow users to swap between the heightMap and colorMap whilst in the editor, so that they can see what the terrain will look like before they press play