Mesh Generation
Added a way to make the generated heightmap into a mesh and extrude it out based upon the noise

Added each vertex into an array and cycle through this array to create an array of triangles, which are used to create a Mesh object and allow it to be displayed as the Unity Mesh API requires

Added a new DrawMode called Mesh, changed noiseMap drawmode to HeightMap

Added a HeightMultiplier to allow the terrain to have more variance in height

Added an animation curve so that the user can dictate at what height the terrain should start getting affected by the increase in height (this is so water isn't raised up like a mountain)

Added a DrawMesh function to allow for the mesh to be created before runtime

Added some comments to remind myself and others of what certain blocks of code are doing

return MeshData instead of the actual Mesh as this is more open ended allowing for the system to be easily upgradeable in the future with threading and chunking for infinite terrain generation if user wish for it. However that is completely out of scope for this project so this will not be worked on.

Filled in what I think are the best presets for each option.

When creating some areas (like a housing area) I will have to draw on the height map before the mesh is created to allow for the mesh to adapt to the change of the heightmap.