Inspector Bug Squash
Squashed several unintended effects of the UI's interacting with each other

Easy now to add on to this if a bit time consuming. I can already see the UI scripts becoming much bigger in scale than the main features of this tool.

I am to inexperienced with Unity Editor at the moment to understand exactly the best way to approach the design of these UI so some research into this after the project is finished would be ideal.