Added Flatshading and Falloff Map
Added a falloff map for the user to tick on or off. When it is on it will surround the landmass generated by water, giving the appearance of an island - Code from Sebastian Lague

Added Flatshading which gives each triangle one normal value, meaning only one shade of colour, as the light acts the same upon the triangle across the whole triangle. Code from Sebastian (adapted to work within my own project)