Added all of the Cellular Automata and Marching squares from Sebastian Lague
Code added to allow for cave generations. This actually looks very cool so this might as well be kept as is as an extra feature. Planned reversal of code still to take place, but will be in addition to this feature. So the user will be able to generate caves but also terrain from the same technique.