House spacing
It has taken a lot of my time to try and sort out the houses in the housing area. Trying to get them in specific places whilst thinking of how to design an open generation where the spacings could vary for each user and choices they would want to make.

Unfortunately I havent been able to implement this and I have spent far too much time trying to do this as evidenced by the lack of commits on this repo. I will endeavor to return to this in the future before completing this module however I will need to move on to the vegetation generation so that there is at least something to show in the upcoming Viva defence