Update to Bug Fix
The last bug fix attempted to solve the problem but somehow made the houses dissapear whenever you exited Unity even if you saved the project.

This is very strange behaivbiour and I don't quite know what is triggering this to happen. The previous bug fix has been reverted to stop the issue of deleting houses and users will for the time being delete houses from the scene and reinitialise the list by commenting it out and back in again