Added Tree spawning
Added code that spawns trees in at 3 random locations and will then spawn trees in around them and scale them down the further out the trees are to simulate age of the tree.

Next step is to slightly shift each trees location so that the trees dont feel boxy as they are spawned on a grid style.

The step after this will look at spawning trees at random locations with a more sparse population and maybe a tertiary stage that is even more sparse than the secondary stage. This should then scale to an attribute able to be set by the user of time of vegetation growth to simulate the growth of these individual forested areas and trees that are created after them.