Updated Tree Generations
Trees now have a random rotation and scale. The rotation is completely random within the range of 0 and 360. The Scale on the other hand is directly linked to the position the tree is in compared to it's ancestor tree. There is a variance of 0.1f scale in Unity for every tree that is the same distance away in the grid. so if the trees average height is suppossed to be a scale of 1.0 then the tree could be as low as 0.95 or as high as 1.05. This along with the rotation allows for all of the trees to look unique and not copied. Also having 11 different trees to use for this helps massively to reduce the appearance of reused assets.

Trees also now spawn a 20x20 grid around them of trees. As the grid gets further out from the ancestor tree the trees have less chance of spawning. This is a good simulation off how trees would spread as a forest ids thinner on the outside than it is on the outside.

Small bug fixed with ancestor trees where wrong trees co-ords were referenced.

I will now be looking at small variances in the positioning of each tree in the grid to make the forests look less boxy