Vegetation Optimisation, variance in positioning on grid and random clusters of trees
Added optimisation to the vegetation spawning looking as much as possible to the DRY technique.

Added variance in the positioning of each tree on the grid reducing the box like impression of the forests and small clusters of trees.

Added small clusters and lone trees to everywhere that is not currently blocked on the noiseMap *based on a chance for spawning*

Added some comments to allow me to know where I am in the script as the script becomes huge.