Flowers, bushes and grass
Added the spawning of other vegetation, although this hasn't quite gone as I had hoped.

I had hoped to keep the standard noiseMapBlocked without the trees although for some reason I cant figure out how to copy the array rather than reference it as I haven't done this in Unity and I am unfamiliar with their methods.

This results in me having to use the noiseMapBlocked array as reference of where the flowers can spawn, which is a shame as they should be allowed to spawn under trees.