Update to Bug Fixing
Managed to add a fixed version of the deleting models button allowing for easy deletion of models before ending a session if the user so wishes. This will mean the user wont have to manually delete each model at the start of their next session.

Whilst this is now fixed an issue has been found with the manual deletion in the project as there are far too many objects now in the scene. This also affects the auto updating method as it is far too slow updating. I will atempt to fix this by adding an if statement so if autoUpdate is on it will not generate the architecture or vegetation. This will drastically improve the speed of this.

I will also try to add in a button the user can press to then generate just the architecture and vegetation on top of the existing terrain so even if the user has random seed on it will keep the terrain the user has selected. The user would also be able to use the generate button if they didn't have random seed selected, as the same terrain would be generated.