Bug Fixing
Added a way for users to press play and generate into the map after it has finished loading. Although this doesn't really allow for users to explore exactly what they have created yet, it is close. Some further options are needed to keep the vegetation and architecture in the exact same spots to allow for this.

Fixed the longest standing bug that I could not figure out for the life of me. It turns out I didn't initialise the list that was affecting having to comment out and then uncommenting out lines of code to initialise the list. I knew this was the problem, initialising the list. However, I didn't realise I should initialise it as I didn't want to inialise the list every time through. THis has been fixed with a try catch. At first the code will attempt to delete from the list and if it cant (because it isn't initialised) it will initialise the list. This works in both editor and play mode in Unity.

Updated some of the code to only allow the generation of the models if autoUpdate and UseRandomSeed are not ticked on. This is to reduce the time taken to load the autoUpdate and move the models in the scene (regenerating them etc.) Users can click the generate models button after clicking the Generate button (when autoUpdate is on) to generate all of the models. This is sometimes glitchy but it works most of the time so I will call this a success for now.