Added Terracing Housing layout
Terracing Housing Layout is usable, however it is noticeably not correct at the moment as resizing of the flattened terrain needs to occur to allow for this layout to look more... normal.

Noticed a bug with the trees and flowers spawning in the Housing area which I thought I had  set this to not be allowed to be spawned in by them.